feat: add OwnersOnly and ParamsChecker components
