+++ packages/app/src/common/utils/object.ts
@@ -1,0 +1,10 @@
+/**
+ * Return true if the object has no property.
+ */
+export const isEmpty = (obj: Record<string, unknown>) => {
+  for (const prop in obj) {
+    if (Object.prototype.hasOwnProperty.call(obj, prop)) return false;
+  }
+
+  return true;
+};
