+++ packages/app/src/components/AuthenticatedOnly.tsx
@@ -1,33 +1,16 @@
-import { useAutoAnimate } from "@formkit/auto-animate/react";
 import type { PropsWithChildren } from "react";
 import React from "react";
 
-import { Alert, AlertTitle } from "../design-system/base/Alert";
 import { useUser } from "@services/apiClient";
 
+/**
+ * Boundary component to check if user is authenticated. If not, it will be redirect to email page.
+ * It not renders children until the auth check is completed.
+ */
 export const AuthenticatedOnly = ({ children, ...delegated }: PropsWithChildren<unknown>) => {
   const { isAuthenticated, loading } = useUser({ redirectTo: "/representation-equilibree/email" });
 
   if (loading || !isAuthenticated) return null;
-
-  return <div {...delegated}>{children}</div>;
-};
-
-export const StaffOnly = ({ children, ...delegated }: PropsWithChildren<unknown>) => {
-  const { user, loading } = useUser({ redirectTo: "/representation-equilibree/email" });
-  const [animationParent] = useAutoAnimate<HTMLDivElement>();
-
-  if (loading || !user) return null;
-
-  if (!user?.staff)
-    return (
-      <div ref={animationParent} style={{ marginBottom: 20 }}>
-        <Alert type="error">
-          <AlertTitle>Contrôle d'accès</AlertTitle>
-          Vous n'avez pas les droits pour accéder à cette page.
-        </Alert>
-      </div>
-    );
 
   return <div {...delegated}>{children}</div>;
 };
